# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a golang web template server.",
        "title": "Golang Web Template API",
        "contact": {
            "name": "API Support",
            "url": "http://mobigen.com",
            "email": "jblim@mobigen.com"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/v1/sample": {
            "post": {
                "description": "새 샘플을 생성한다",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sample"
                ],
                "summary": "Create sample",
                "parameters": [
                    {
                        "description": "생성할 샘플",
                        "name": "sample",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Sample"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "생성된 샘플",
                        "schema": {
                            "$ref": "#/definitions/HTTPResponse-Sample"
                        }
                    },
                    "400": {
                        "description": "잘못된 요청",
                        "schema": {
                            "$ref": "#/definitions/HTTPResponse-any"
                        }
                    },
                    "500": {
                        "description": "서버 오류",
                        "schema": {
                            "$ref": "#/definitions/HTTPResponse-any"
                        }
                    }
                }
            }
        },
        "/api/v1/sample/update": {
            "post": {
                "description": "샘플을 수정한다",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sample"
                ],
                "summary": "Update sample",
                "parameters": [
                    {
                        "description": "수정할 샘플 (ID 필수)",
                        "name": "sample",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Sample"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "수정된 샘플",
                        "schema": {
                            "$ref": "#/definitions/HTTPResponse-Sample"
                        }
                    },
                    "400": {
                        "description": "잘못된 요청",
                        "schema": {
                            "$ref": "#/definitions/HTTPResponse-any"
                        }
                    },
                    "500": {
                        "description": "서버 오류",
                        "schema": {
                            "$ref": "#/definitions/HTTPResponse-any"
                        }
                    }
                }
            }
        },
        "/api/v1/sample/{id}": {
            "get": {
                "description": "ID로 샘플을 조회한다",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sample"
                ],
                "summary": "Get sample by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Sample ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "샘플",
                        "schema": {
                            "$ref": "#/definitions/HTTPResponse-Sample"
                        }
                    },
                    "400": {
                        "description": "잘못된 파라미터",
                        "schema": {
                            "$ref": "#/definitions/HTTPResponse-any"
                        }
                    },
                    "404": {
                        "description": "샘플 없음",
                        "schema": {
                            "$ref": "#/definitions/HTTPResponse-any"
                        }
                    },
                    "500": {
                        "description": "서버 오류",
                        "schema": {
                            "$ref": "#/definitions/HTTPResponse-any"
                        }
                    }
                }
            },
            "delete": {
                "description": "ID로 샘플을 삭제한다",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sample"
                ],
                "summary": "Delete sample",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Sample ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "삭제된 샘플",
                        "schema": {
                            "$ref": "#/definitions/HTTPResponse-Sample"
                        }
                    },
                    "400": {
                        "description": "잘못된 파라미터",
                        "schema": {
                            "$ref": "#/definitions/HTTPResponse-any"
                        }
                    },
                    "404": {
                        "description": "샘플 없음",
                        "schema": {
                            "$ref": "#/definitions/HTTPResponse-any"
                        }
                    },
                    "500": {
                        "description": "서버 오류",
                        "schema": {
                            "$ref": "#/definitions/HTTPResponse-any"
                        }
                    }
                }
            }
        },
        "/api/v1/samples": {
            "get": {
                "description": "전체 샘플 목록을 반환한다",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sample"
                ],
                "summary": "Get all samples",
                "responses": {
                    "200": {
                        "description": "샘플 목록",
                        "schema": {
                            "$ref": "#/definitions/HTTPResponse-array_Sample"
                        }
                    },
                    "500": {
                        "description": "서버 오류",
                        "schema": {
                            "$ref": "#/definitions/HTTPResponse-any"
                        }
                    }
                }
            }
        },
        "/version": {
            "get": {
                "description": "get server version info",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "version"
                ],
                "summary": "Get Server Version",
                "responses": {
                    "200": {
                        "description": "app info(name, version, hash)",
                        "schema": {
                            "$ref": "#/definitions/HTTPResponse-VersionInfo"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "HTTPResponse-Sample": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/Sample"
                },
                "isSuccess": {
                    "type": "boolean"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "HTTPResponse-VersionInfo": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/VersionInfo"
                },
                "isSuccess": {
                    "type": "boolean"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "HTTPResponse-any": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "isSuccess": {
                    "type": "boolean"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "HTTPResponse-array_Sample": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Sample"
                    }
                },
                "isSuccess": {
                    "type": "boolean"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "Sample": {
            "type": "object",
            "properties": {
                "createAt": {
                    "type": "integer"
                },
                "desc": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "VersionInfo": {
            "type": "object",
            "properties": {
                "buildHash": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        }
    }
}